# [PERFORMANCE] [v1.x] Layer normalization code from Marian for CPU