@kpuatamazon 2020-11-30
Lint broken?  
@mxnet-bot run ci [sanity]

```
[2020-11-30T17:51:32.903Z] [91m+ pip3 install -r /work/requirements
[2020-11-30T17:51:33.157Z] [0m[91mDEPRECATION: Python 3.5 reached the end of its life on September 13th, 2020. Please upgrade your Python as Python 3.5 is no longer maintained. pip 21.0 will drop support for Python 3.5 in January 2021. pip 21.0 will remove support for this functionality.
[2020-11-30T17:51:35.665Z] [0mCollecting astroid==2.3.3
[2020-11-30T17:51:35.665Z]   Downloading astroid-2.3.3-py3-none-any.whl (205 kB)
[2020-11-30T17:51:35.665Z] 
[2020-11-30T17:51:35.665Z] The conflict is caused by:
[2020-11-30T17:51:35.665Z]     The user requested six==1.11.0
[2020-11-30T17:51:35.665Z]     astroid 2.3.3 depends on six~=1.12
[2020-11-30T17:51:35.665Z] 
[2020-11-30T17:51:35.665Z] To fix this you could try to:
[2020-11-30T17:51:35.665Z] 1. loosen the range of package versions you've specified
[2020-11-30T17:51:35.665Z] 2. remove package versions to allow pip attempt to solve the dependency conflict
[2020-11-30T17:51:35.665Z] 
[2020-11-30T17:51:35.665Z] [91mERROR: Cannot install -r /work/requirements (line 31) and six==1.11.0 because these package versions have conflicting dependencies.
[2020-11-30T17:51:35.665Z] ERROR: ResolutionImpossible: for help visit https://pip.pypa.io/en/latest/user_guide/#fixing-conflicting-dependencies
[2020-11-30T17:51:36.222Z] [0mThe command '/bin/sh -c /work/ubuntu_python.sh' returned a non-zero code: 1
[2020-11-30T17:51:36.222Z] Traceback (most recent call last):
[2020-11-30T17:51:36.222Z]   File "ci/build.py", line 456, in <module>
[2020-11-30T17:51:36.222Z]     sys.exit(main())
[2020-11-30T17:51:36.222Z]   File "ci/build.py", line 366, in main
[2020-11-30T17:51:36.222Z]     cache_intermediate=args.cache_intermediate)
[2020-11-30T17:51:36.222Z]   File "ci/build.py", line 114, in build_docker
[2020-11-30T17:51:36.222Z]     run_cmd()
[2020-11-30T17:51:36.222Z]   File "/home/jenkins_slave/workspace/sanity-lint/ci/util.py", line 84, in f_retry
[2020-11-30T17:51:36.222Z]     return f(*args, **kwargs)
[2020-11-30T17:51:36.222Z]   File "ci/build.py", line 112, in run_cmd
[2020-11-30T17:51:36.222Z]     check_call(cmd)
[2020-11-30T17:51:36.222Z]   File "/usr/lib/python3.6/subprocess.py", line 311, in check_call
[2020-11-30T17:51:36.222Z]     raise CalledProcessError(retcode, cmd)
[2020-11-30T17:51:36.222Z] subprocess.CalledProcessError: Command '['docker', 'build', '-f', 'docker/Dockerfile.build.ubuntu_cpu', '--build-arg', 'USER_ID=1001', '--build-arg', 'GROUP_ID=1001', '--cache-from', 'mxnetci/build.ubuntu_cpu', '-t', 'mxnetci/build.ubuntu_cpu', 'docker']' returned non-zero exit status 1.
script returned exit code 1
```