@kpuatamazon 2020-12-04
@mxnet-bot run ci [sanity]

Maybe #19604 fixed lint?