@kpuatamazon 2020-12-07
@mxnet-bot run ci [sanity]

Looks like the gcc-8 PPA has disappeared #19636
```
[2020-12-07T13:17:24.044Z] Step 24/38 : RUN /work/ubuntu_gcc8.sh
[2020-12-07T13:17:24.044Z]  ---> Running in b5115b7afd64
[2020-12-07T13:17:26.557Z] Cannot add PPA: 'ppa:~jonathonf/ubuntu/gcc-8.0'.
[2020-12-07T13:17:26.557Z] The user named '~jonathonf' has no PPA named 'ubuntu/gcc-8.0'
[2020-12-07T13:17:26.557Z] Please choose from the following available PPAs:
[2020-12-07T13:17:26.557Z]  * 'ansible':  Ansible
[2020-12-07T13:17:26.557Z]  * 'atslang':  ATS2 programming language
[2020-12-07T13:17:26.557Z]  * 'backports':  Backport collection
[2020-12-07T13:17:26.557Z]  * 'barrier':  Barrier
[2020-12-07T13:17:26.557Z]  * 'bazel':  Bazel
[2020-12-07T13:17:26.557Z]  * 'byobu':  Byobu
[2020-12-07T13:17:26.557Z]  * 'calibre':  Calibre
[2020-12-07T13:17:26.557Z]  * 'clementine':  Clementine
[2020-12-07T13:17:26.557Z]  * 'compton':  Compton
[2020-12-07T13:17:26.557Z]  * 'cpan':  Packaged CPAN modules
[2020-12-07T13:17:26.557Z]  * 'debhelper':  debhelper
[2020-12-07T13:17:26.557Z]  * 'debhelper-10':  debhelper 10
[2020-12-07T13:17:26.557Z]  * 'debhelper-11':  debhelper 11
[2020-12-07T13:17:26.557Z]  * 'debhelper-12':  debhelper 12
[2020-12-07T13:17:26.557Z]  * 'debhelper-trusty':  debhelper for trusty
[2020-12-07T13:17:29.078Z] Cannot add PPA: 'ppa:~jonathonf/ubuntu/gcc-7.3'.
[2020-12-07T13:17:29.078Z] The user named '~jonathonf' has no PPA named 'ubuntu/gcc-7.3'
[2020-12-07T13:17:29.078Z] Please choose from the following available PPAs:
[2020-12-07T13:17:29.078Z]  * 'ansible':  Ansible
[2020-12-07T13:17:29.078Z]  * 'atslang':  ATS2 programming language
[2020-12-07T13:17:29.078Z]  * 'backports':  Backport collection
[2020-12-07T13:17:29.078Z]  * 'barrier':  Barrier
[2020-12-07T13:17:29.078Z]  * 'bazel':  Bazel
[2020-12-07T13:17:29.078Z]  * 'byobu':  Byobu
[2020-12-07T13:17:29.078Z]  * 'calibre':  Calibre
[2020-12-07T13:17:29.078Z]  * 'clementine':  Clementine
[2020-12-07T13:17:29.078Z]  * 'compton':  Compton
[2020-12-07T13:17:29.078Z]  * 'cpan':  Packaged CPAN modules
[2020-12-07T13:17:29.078Z]  * 'debhelper':  debhelper
[2020-12-07T13:17:29.078Z]  * 'debhelper-10':  debhelper 10
[2020-12-07T13:17:29.078Z]  * 'debhelper-11':  debhelper 11
[2020-12-07T13:17:29.078Z]  * 'debhelper-12':  debhelper 12
[2020-12-07T13:17:29.078Z]  * 'debhelper-trusty':  debhelper for trusty
[2020-12-07T13:17:29.078Z] Hit:1 https://apt.llvm.org/xenial llvm-toolchain-xenial-3.9 InRelease
[2020-12-07T13:17:29.078Z] Hit:2 https://apt.llvm.org/xenial llvm-toolchain-xenial-6.0 InRelease
[2020-12-07T13:17:29.078Z] Ign:3 http://cran.rstudio.com/bin/linux/ubuntu trusty/ InRelease
[2020-12-07T13:17:29.078Z] Hit:4 http://cran.rstudio.com/bin/linux/ubuntu trusty/ Release
[2020-12-07T13:17:29.078Z] Hit:5 http://security.ubuntu.com/ubuntu xenial-security InRelease
[2020-12-07T13:17:29.078Z] Hit:6 http://archive.ubuntu.com/ubuntu xenial InRelease
[2020-12-07T13:17:29.078Z] Hit:8 http://archive.ubuntu.com/ubuntu xenial-updates InRelease
[2020-12-07T13:17:29.333Z] Hit:9 http://archive.ubuntu.com/ubuntu xenial-backports InRelease
[2020-12-07T13:17:29.894Z] Reading package lists...
[2020-12-07T13:17:30.822Z] Reading package lists...
[2020-12-07T13:17:30.822Z] Building dependency tree...
[2020-12-07T13:17:30.822Z] Reading state information...
[2020-12-07T13:17:30.822Z] �[91mE: Unable to locate package gcc-8
[2020-12-07T13:17:31.077Z] �[0mThe command '/bin/sh -c /work/ubuntu_gcc8.sh' returned a non-zero code: 100
[2020-12-07T13:17:31.077Z] Traceback (most recent call last):
[2020-12-07T13:17:31.077Z]   File "ci/build.py", line 456, in <module>
[2020-12-07T13:17:31.077Z]     sys.exit(main())
[2020-12-07T13:17:31.077Z]   File "ci/build.py", line 366, in main
[2020-12-07T13:17:31.077Z]     cache_intermediate=args.cache_intermediate)
[2020-12-07T13:17:31.077Z]   File "ci/build.py", line 114, in build_docker
[2020-12-07T13:17:31.077Z]     run_cmd()
[2020-12-07T13:17:31.077Z]   File "/home/jenkins_slave/workspace/sanity-lint/ci/util.py", line 84, in f_retry
[2020-12-07T13:17:31.077Z]     return f(*args, **kwargs)
[2020-12-07T13:17:31.077Z]   File "ci/build.py", line 112, in run_cmd
[2020-12-07T13:17:31.077Z]     check_call(cmd)
[2020-12-07T13:17:31.077Z]   File "/usr/lib/python3.6/subprocess.py", line 311, in check_call
[2020-12-07T13:17:31.077Z]     raise CalledProcessError(retcode, cmd)
[2020-12-07T13:17:31.077Z] subprocess.CalledProcessError: Command '['docker', 'build', '-f', 'docker/Dockerfile.build.ubuntu_cpu', '--build-arg', 'USER_ID=1001', '--build-arg', 'GROUP_ID=1001', '--cache-from', 'mxnetci/build.ubuntu_cpu', '-t', 'mxnetci/build.ubuntu_cpu', 'docker']' returned non-zero exit status 100.
script returned exit code 1\
```